new try

+++ treemap.c
@@ -39,7 +39,7 @@ TreeNode * createTreeNode(void* key, void * value) {
 TreeMap * createTreeMap(int (*lower_than) (void* key1, void* key2)) {
         TreeMap * new = (TreeMap *)malloc(sizeof(TreeMap));
         new->lower_than = lower_than;
-        printf("%d",new->lower_than);
+        printf("%d",(int)new->lower_than);
         return NULL;
 
 }
